Initial attempt to add web detection!
