+++ examples/resizeImages/index.html
@@ -85,8 +85,8 @@ setTimeout(function() {
         return true;
     },
 
-    // mobify.js + main executable (combined for improved performance)
-    '/mobifyjs/examples/resizeImages/mobify+main.js'
+    '/build/mobify.js',
+    '/mobifyjs/examples/resizeImages/main.js'
 );
 </script>
 <!-- END MOBIFY TAG -->
