--- examples/resizeImages/mobify+main.js
@@ -1,24 +1,0 @@
-// mobify.js
-(function(){var e,t,n;(function(r){function d(e,t){return h.call(e,t)}function v(e,t){var n,r,i,s,o,u,a,f,c,h,p=t&&t.split("/"),d=l.map,v=d&&d["*"]||{};if(e&&e.charAt(0)===".")if(t){p=p.slice(0,p.length-1),e=p.concat(e.split("/"));for(f=0;f<e.length;f+=1){h=e[f];if(h===".")e.splice(f,1),f-=1;else if(h===".."){if(f===1&&(e[2]===".."||e[0]===".."))break;f>0&&(e.splice(f-1,2),f-=2)}}e=e.join("/")}else e.indexOf("./")===0&&(e=e.substring(2));if((p||v)&&d){n=e.split("/");for(f=n.length;f>0;f-=1){r=n.slice(0,f).join("/");if(p)for(c=p.length;c>0;c-=1){i=d[p.slice(0,c).join("/")];if(i){i=i[r];if(i){s=i,o=f;break}}}if(s)break;!u&&v&&v[r]&&(u=v[r],a=f)}!s&&u&&(s=u,o=a),s&&(n.splice(0,o,s),e=n.join("/"))}return e}function m(e,t){return function(){return s.apply(r,p.call(arguments,0).concat([e,t]))}}function g(e){return function(t){return v(t,e)}}function y(e){return function(t){a[e]=t}}function b(e){if(d(f,e)){var t=f[e];delete f[e],c[e]=!0,i.apply(r,t)}if(!d(a,e)&&!d(c,e))throw new Error("No "+e);return a[e]}function w(e){var t,n=e?e.indexOf("!"):-1;return n>-1&&(t=e.substring(0,n),e=e.substring(n+1,e.length)),[t,e]}function E(e){return function(){return l&&l.config&&l.config[e]||{}}}var i,s,o,u,a={},f={},l={},c={},h=Object.prototype.hasOwnProperty,p=[].slice;o=function(e,t){var n,r=w(e),i=r[0];return e=r[1],i&&(i=v(i,t),n=b(i)),i?n&&n.normalize?e=n.normalize(e,g(t)):e=v(e,t):(e=v(e,t),r=w(e),i=r[0],e=r[1],i&&(n=b(i))),{f:i?i+"!"+e:e,n:e,pr:i,p:n}},u={require:function(e){return m(e)},exports:function(e){var t=a[e];return typeof t!="undefined"?t:a[e]={}},module:function(e){return{id:e,uri:"",exports:a[e],config:E(e)}}},i=function(e,t,n,i){var s,l,h,p,v,g=[],w;i=i||e;if(typeof n=="function"){t=!t.length&&n.length?["require","exports","module"]:t;for(v=0;v<t.length;v+=1){p=o(t[v],i),l=p.f;if(l==="require")g[v]=u.require(e);else if(l==="exports")g[v]=u.exports(e),w=!0;else if(l==="module")s=g[v]=u.module(e);else if(d(a,l)||d(f,l)||d(c,l))g[v]=b(l);else{if(!p.p)throw new Error(e+" missing "+l);p.p.load(p.n,m(i,!0),y(l),{}),g[v]=a[l]}}h=n.apply(a[e],g);if(e)if(s&&s.exports!==r&&s.exports!==a[e])a[e]=s.exports;else if(h!==r||!w)a[e]=h}else e&&(a[e]=n)},e=t=s=function(e,t,n,a,f){return typeof e=="string"?u[e]?u[e](t):b(o(e,t).f):(e.splice||(l=e,t.splice?(e=t,t=n,n=null):e=r),t=t||function(){},typeof n=="function"&&(n=a,a=f),a?i(r,e,t,n):setTimeout(function(){i(r,e,t,n)},4),s)},s.config=function(e){return l=e,l.deps&&s(l.deps,l.callback),s},n=function(e,t,n){t.splice||(n=t,t=[]),!d(a,e)&&!d(f,e)&&(f[e]=[e,t,n])},n.amd={jQuery:!0}})(),n("almond",function(){}),n("utils",[],function(){var e={};return e.extend=function(e){return[].slice.call(arguments,1).forEach(function(t){for(var n in t)t[n]!==undefined&&(e[n]=t[n])}),e},e.keys=function(e){var t=[];for(var n in e)e.hasOwnProperty(n)&&t.push(n);return t},e.values=function(e){var t=[];for(var n in e)e.hasOwnProperty(n)&&t.push(e[n]);return t},e.clone=function(e){var t={};for(var n in e)e.hasOwnProperty(n)&&(t[n]=e[n]);return t},e.outerHTML=function(e){var t=document.createElement("div");t.appendChild(e.cloneNode(!0));var n=t.innerHTML;return t=null,n},e.removeBySelector=function(t,n){n=n||document;var r=n.querySelectorAll(t);return e.removeElements(r,n)},e.removeElements=function(e,t){t=t||document;for(var n=0,r=e.length;n<r;n++){var i=e[n];i.parentNode.removeChild(i)}return e},e}),n("capture",["utils"],function(e){function l(e){return e.nodeName.toLowerCase()}function c(e){return e.replace('"',"&quot;")}function h(t){return t?[].map.call(t.childNodes,function(t){var n=l(t);return n=="#comment"?"<!--"+t.textContent+"-->":n=="plaintext"?t.textContent:n=="script"&&(/mobify/.test(t.src)||/mobify/i.test(t.textContent))?"":t.outerHTML||t.nodeValue||e.outerHTML(t)}).join(""):""}window.Mobify&&!window.Mobify.capturing&&document.getElementsByTagName("plaintext").length&&(window.Mobify.capturing=!0);var t=/(<script[\s\S]*?>)/gi,n={style:' media="mobify-media"',script:' type="text/mobify-script"'},r=new RegExp(e.values(n).join("|"),"g"),i={img:["src"],iframe:["src"],script:["src","type"],link:["href"],style:["media"]},s=new RegExp("<("+e.keys(i).join("|")+")([\\s\\S]*?)>","gi"),o={},u={};for(var a in i){if(!i.hasOwnProperty(a))continue;var f=i[a];f.forEach(function(e){u[e]=!0}),o[a]=new RegExp("\\s+((?:"+f.join("|")+")\\s*=\\s*(?:('|\")[\\s\\S]+?\\2))","gi")}var p=document.createElement("div"),d=function(t,n){this.doc=t,this.prefix=n||"x-";var r=this.createDocumentFragmentsStrings();e.extend(this,r);var i=this.createDocumentFragments();e.extend(this,i)},v=d.init=function(e,t,n){var t=t||document,r=function(e,t,n){var r=new d(t,n);e(r)};if(/complete|interactive|loaded/.test(t.readyState))r(e,t,n);else{var i=!1;t.addEventListener("readystatechange",function(){i||(i=!0,r(e,t,n))},!1)}};return d.cloneAttributes=function(e,t){var n=e.match(/^<(\w+)([\s\S]*)$/i);return p.innerHTML="<div"+n[2],[].forEach.call(p.firstChild.attributes,function(e){try{t.setAttribute(e.nodeName,e.nodeValue)}catch(n){console.error("Error copying attributes while capturing: ",n)}}),t},d.disable=function(e,r){var i=this,u=function(){return function(e,t,i){return lowercaseTagName=t.toLowerCase(),result="<"+lowercaseTagName+(n[lowercaseTagName]||"")+i.replace(o[lowercaseTagName]," "+r+"$1")+">"}}(),a=/(<!--[\s\S]*?-->)|(?=<\/script)/i,f=e.split(a),l=f.map(function(e){var n;return e?/^<!--/.test(e)?e:(n=e.split(t),n[0]=n[0].replace(s,u),n[1]&&(n[1]=n[1].replace(s,u)),n):""});return[].concat.apply([],l).join("")},d.enable=function(t,n){var i=new RegExp("\\s"+n+"("+e.keys(u).join("|")+")","gi");return t.replace(i," $1").replace(r,"")},d.openTag=function(e){if(!e)return"";e.length&&(e=e[0]);var t=[];return[].forEach.call(e.attributes,function(e){t.push(" ",e.name,'="',c(e.value),'"')}),"<"+l(e)+t.join("")+">"},d.prototype.getDoctype=function(){var e=this.doc.doctype||[].filter.call(this.doc.childNodes,function(e){return e.nodeType==Node.DOCUMENT_TYPE_NODE})[0];return e?"<!DOCTYPE HTML"+(e.publicId?' PUBLIC "'+e.publicId+'"':"")+(e.systemId?' "'+e.systemId+'"':"")+">":""},d.prototype.createDocumentFragmentsStrings=function(){var e=this.doc,t=e.getElementsByTagName("head")[0]||e.createElement("head"),n=e.getElementsByTagName("body")[0]||e.createElement("body"),r=e.getElementsByTagName("html")[0];captured={doctype:this.getDoctype(),htmlOpenTag:d.openTag(r),headOpenTag:d.openTag(t),bodyOpenTag:d.openTag(n),headContent:h(t),bodyContent:h(n)},captured.all=function(e){return this.doctype+this.htmlOpenTag+this.headOpenTag+(e||"")+this.headContent+this.bodyContent};var i=/<!--(?:[\s\S]*?)-->|(<\/head\s*>|<body[\s\S]*$)/gi,s=captured.bodyContent=captured.headContent+captured.bodyContent;captured.headContent="";for(var o;o=i.exec(s);o){if(!o[1])continue;if(o[1][1]!="/"){captured.headContent=captured.head||s.slice(0,o.index),captured.bodyContent=o[0];var u=/^((?:[^>'"]*|'[^']*?'|"[^"]*?")*>)([\s\S]*)$/.exec(captured.bodyContent);u&&(captured.bodyOpenTag=u[1],captured.bodyContent=u[2]);break}captured.headContent=s.slice(0,o.index),captured.bodyContent=s.slice(o.index+o[1].length)}return captured},d.prototype.restore=function(){var e=this,t=e.doc,n=function(){t.removeEventListener("DOMContentLoaded",n,!1),setTimeout(function(){t.open(),t.write(e.all()),t.close()},15)};/complete|interactive|loaded/.test(t.readyState)?n():t.addEventListener("DOMContentLoaded",n,!1)},d.prototype.setElementContentFromString=function(e,t){var n=this.doc;for(p.innerHTML=t;p.firstChild;e.appendChild(p.firstChild));},d.prototype.createDocumentFragments=function(){var e={},t=e.capturedDoc=document.implementation.createHTMLDocument(""),n=e.htmlEl=t.documentElement,r=e.headEl=n.firstChild,i=e.bodyEl=n.lastChild;d.cloneAttributes(this.htmlOpenTag,n),d.cloneAttributes(this.headOpenTag,r),d.cloneAttributes(this.bodyOpenTag,i),i.innerHTML=d.disable(this.bodyContent,this.prefix);var s=d.disable(this.headContent,this.prefix);try{r.innerHTML=s}catch(o){var u=r.getElementsByTagName("title")[0];u&&r.removeChild(u),this.setElementContentFromString(r,s)}return n.appendChild(r),n.appendChild(i),e},d.prototype.escapedHTMLString=function(){var t=this.capturedDoc,n=d.enable(t.documentElement.outerHTML||e.outerHTML(t.documentElement),this.prefix),r=this.doctype+n;return r},d.prototype.render=function(e){var t;e?t=d.enable(e):t=this.escapedHTMLString();var n=this.doc;window.Mobify&&(window.Mobify.capturing=!1),setTimeout(function(){n.open("text/html","replace"),n.write(t),n.close()})},d.prototype.getCapturedDoc=function(e){return this.capturedDoc},d.prototype.insertMobifyScripts=function(){var e=this.capturedDoc,t=document.getElementById("mobify-js");t||(t=document.getElementsByTagName("script")[0],t.id="mobify-js",t.setAttribute("class","mobify"));var n=e.importNode(t,!1),r=this.headEl;r.insertBefore(n,r.firstChild);var i=document.getElementById("mobify-js-main");if(i){var s=e.importNode(i,!1);this.bodyEl.appendChild(s)}},d.prototype.renderCapturedDoc=function(e){var t=this.capturedDoc;this.insertMobifyScripts();if(window.Mobify&&window.Mobify.points){var n=this.bodyEl,r=t.createElement("div");r.id="mobify-point",r.setAttribute("style","display: none;"),r.innerHTML=window.Mobify.points[0],n.insertBefore(r,n.firstChild)}this.render()},d}),n("resizeImages",["utils"],function(e){function s(e){function t(t){var n=e||1;return t.width=Math.round(t.width*n),t.height=Math.round(t.height*n),t}var n=navigator.userAgent.match(/iphone|ipod|ipad/i),r=(navigator.userAgent.match(/android (\d)/i)||{})[1],i={width:window.outerWidth,height:window.outerHeight};if(!n)return r>3?t(i):i;var s=window.orientation%180;return s?(i.height=screen.width,i.width=screen.height):(i.width=screen.width,i.height=screen.height),t(i)}var t={},n=document.createElement("a"),r=/^https?/,i="//ir0.mobify.com",o=t.getImageURL=function(t,n){var r=e.clone(u);n&&e.extend(r,n);var s=[i];if(r.projectName){var o="project-"+r.projectName;s.push(o)}return n.cacheHours&&s.push("c"+n.cacheHours),r.format&&s.push(n.format+(n.quality||"")),r.maxWidth&&(s.push(n.maxWidth),r.maxHeight&&s.push(n.maxHeight)),s.push(t),s.join("/")};t.resize=function(t,i){var a=e.clone(u);i&&e.extend(a,i);var f=a.devicePixelRatio||window.devicePixelRatio,l=s(f),c=a.maxWidth||l.width,h=a.maxHeight||l.height;f&&a.maxWidth&&(c*=f,a.maxHeight&&(h*=f)),a.maxWidth=Math.ceil(c),a.maxHeight=Math.ceil(h);var p;for(var d=0;d<t.length;d++){var v=t[d];if(p=v.getAttribute(a.attribute)){n.href=p;var m=n.href;r.test(m)&&v.setAttribute(a.attribute,o(m,a))}}return t};var u={projectName:"oss-"+location.hostname.replace(/[^\w]/g,"-"),attribute:"x-src"};return t}),n("jazzcat",["utils","capture"],function(e,t){var n=window.Jazzcat={},r=function(e,t){var n=f[e.split("#")[0]];return n&&t&&(n.lastUsed=Date.now(),n.useCount=n.useCount++||1),n},i=function(e,t){f[e]=t},s=function(){var e=localStorage.getItem(a),t;if(e===null)return;try{e=JSON.parse(e)}catch(n){return}for(t in e)e.hasOwnProperty(t)&&!d.utils.isStale(e[t])&&i(t,e[t])},o=function(e){var t={},n,r=10,i,s,o=9007199254740991,u;for(i in f)f.hasOwnProperty(i)&&(t[i]=f[i]);(function l(){setTimeout(function(){try{s=JSON.stringify(t)}catch(i){e&&e(i);return}try{localStorage.setItem(a,s)}catch(i){if(!--r){e&&e(i);return}for(var f in t){if(!t.hasOwnProperty(f))continue;n=t[f];if(!n.lastUsed){u=f,o=0;break}n.lastUsed<=o&&(u=f,o=n.lastUsed)}delete t[u],l();return}e&&e()},0)})()},u=function(e){f=e||{}},a="Mobify-Combo-Cache-v1.0",f={};n.httpCache={get:r,set:i,load:s,save:o,reset:u,cache:f};var l=/^\s*(public|private|no-cache|no-store)\s*$/,c=/^\s*(max-age)\s*=\s*(\d+)\s*$/,h=function(e){var t={},n;return e.split(",").forEach(function(e){if(n=l.exec(e))t[n[1]]=!0;else if(n=c.exec(e))t[n[1]]=parseInt(n[2])}),t},p=n.httpCache.utils={dataURI:function(e){var t=e.headers["content-type"]||"application/x-javascript";return"data:"+t+(e.text?","+encodeURIComponent(e.body):";base64,"+e.body)},isStale:function(e){var t=e.headers||{},n=t["cache-control"],r=Date.now(),i,s;if(n&&(i=Date.parse(t.date))){n=h(n);if(n["max-age"]&&!n["private"]&&!n["no-store"]&&!n["no-cache"])return r>i+n["max-age"]*1e3}return(s=Date.parse(t.expires))?r>s:!0}},d=n.httpCache,v=document.createElement("a");n.combineScripts=function(t,n){var r,i;n?r=e.extend(m,n):r=m,i=function(){var e="modernizr";try{return localStorage.setItem(e,e),localStorage.removeItem(e),!0}catch(t){return!1}};if(!t.length||!i()||!window.JSON)return t;var s,o,u,a;d.load();for(o=0,u=t.length;o<u;o++){var f=t[o];if(!f.hasAttribute(r.attribute))continue;v.href=f.getAttribute(r.attribute),s=v.href,f.removeAttribute(r.attribute),a=!!d.get(s),f.innerHTML=a+","+r.execCallback+"('"+s+"',"+!!r.forceDataURI+");"}return t};var m=n.combineScripts.defaults={selector:"script",attribute:"x-src",proto:"//",host:"jazzcat.mobify.com",endpoint:"jsonp",execCallback:"Jazzcat.combo.exec",loadCallback:"Jazzcat.combo.load",projectName:""};n.combo={exec:function(e,t){var n=d.get(e,!0),r;n?(r='data-orig-src="'+e+'"',t?r+=' src="'+d.utils.dataURI(n)+'">':r+=">"+n.body.replace(/(<\/scr)(ipt\s*>)/ig,"$1\\$2")):r='src="'+e+'">',document.write("<script "+r+"</script>")},load:function(e){var t,n,r,i=!1;d.load();if(!e)return;for(n=0,r=e.length;n<r;n++)t=e[n],t.status=="ready"&&(i=!0,d.set(encodeURI(t.url),t));i&&d.save()},getLoaderScript:function(e,t){var r=document.createElement("script");return e.length?r.src=n.getURL(e,t):r.innerHTML=t+"();",r}},n.getURL=function(e,t){return m.proto+m.host+(m.projectName?"/project-"+m.projectName:"")+"/"+m.endpoint+"/"+t+"/"+n.JSONURIencode(e.slice().sort())},n.JSONURIencode=function(e){return encodeURIComponent(JSON.stringify(e))};var g=t.enable,y=new RegExp("<script[^>]*?>(true|false),"+m.execCallback.replace(/\./g,"\\.")+"\\('([\\s\\S]*?)'\\,(true|false)\\);<\\/script","gi");return t.enable=function(){var e,r,i=-1,s=[],o=g.apply(t,arguments);while(e=y.exec(o))i==-1&&(i=e.index),e[1]==="false"&&s.push(e[2]);return i==-1?o:(r=n.combo.getLoaderScript(s,m.loadCallback),o.substr(0,i)+r.outerHTML+o.substr(i))},n}),n("unblockify",["utils","capture"],function(e,t){var n={};return n.moveScripts=function(t,n){e.removeElements(t,n);for(var r=0,i=t.length;r<i;r++){var s=t[r];n.body.appendChild(s)}},n.unblock=function(e){var r=t.prototype.insertMobifyScripts;t.prototype.insertMobifyScripts=function(){r.call(this);var t=this.capturedDoc;n.moveScripts(e,t)}},n}),t(["utils","capture","resizeImages","jazzcat","unblockify"],function(e,t,n,r,i){var s=window.Mobify=window.Mobify||{};return s.Utils=e,s.Capture=t,s.ResizeImages=n,s.Jazzcat=r,s.Unblockify=i,s.api="2.0",s},undefined,!0),n("mobify-library",function(){})})();
-
-// main executable
-var capturing = window.Mobify && window.Mobify.capturing || false;
-
-if (capturing) {
-    console.log("Executing main during capturing phase!")
-
-    // Grab reference to a newly created document
-    Mobify.Capture.init(function(capture){
-        var capturedDoc = capture.capturedDoc;
-
-        // Resize images using Mobify Image Resizer
-        var images = capturedDoc.querySelectorAll('img');
-        Mobify.ResizeImages.resize( images, {
-            maxWidth: 320   
-        });
-
-        // Render source DOM to document
-        capture.renderCapturedDoc();
-    });
-
-}
